Split runtime host services and guard boundary
